feat: 遷移登入群組（login/login-email/login-verify/signin-email）至 App Router
- 四頁移入 (bare) route group（無 Header），server shell（metadata）+ client 元件
- next/router → next/navigation（皆為 router.push，API 相容）
- login：next/head → Metadata，並移除因此多出的多餘 Fragment
- 登入 API（/api/auth/*）暫留 pages/api，client fetch 照常運作
- 驗證：next build 綠燈、四頁皆靜態預渲染（○）
