+++ src/app/(bare)/login-email/login-email-client.tsx
@@ -1,11 +1,13 @@
+'use client';
+
 import { useState } from 'react';
 import { useForm } from 'react-hook-form';
 import { zodResolver } from '@hookform/resolvers/zod';
 import * as z from 'zod';
 import Image from 'next/image';
 import Link from 'next/link';
-import { X, Lock } from 'lucide-react';
-import { useRouter } from 'next/router';
+import { X } from 'lucide-react';
+import { useRouter } from 'next/navigation';
 
 import { Button } from '@/components/ui/button';
 import {
@@ -16,22 +18,33 @@ import {
   FormMessage,
 } from '@/components/ui/form';
 import { Input } from '@/components/ui/input';
+import { Checkbox } from '@/components/ui/checkbox';
+import { registerWithEmail } from '@/services/auth';
 import { COMMON_TEXTS } from '@/lib/constants/messages';
 import { VALIDATION_MESSAGES } from '@/lib/constants/validation';
 
 // 定義表單驗證規則
 const formSchema = z.object({
+  name: z.string().min(1, VALIDATION_MESSAGES.REQUIRED_YOUR_NAME),
   email: z.string().email(VALIDATION_MESSAGES.INVALID_EMAIL),
-  password: z.string().min(1, VALIDATION_MESSAGES.REQUIRED_PASSWORD),
+  password: z
+    .string()
+    .min(8, '密碼至少需要8個字元')
+    .regex(/[A-Z]/, '密碼需要至少一個大寫字母')
+    .regex(/[0-9]/, '密碼需要至少一個數字')
+    .regex(/^[a-zA-Z0-9]+$/, '密碼只能包含英文字母和數字'),
+  agreement: z.boolean().refine((val) => val === true, {
+    message: '您必須同意條款才能繼續',
+  }),
 });
 
 // 定義表單資料型別
 type FormData = z.infer<typeof formSchema>;
 
 /**
- * 電子郵件登入頁面元件
+ * 電子郵件註冊頁面元件
  */
-export default function SignInWithEmail() {
+export function LoginEmailClient() {
   const router = useRouter();
   const [submitting, setSubmitting] = useState(false);
   const [errorMessage, setErrorMessage] = useState('');
@@ -40,8 +53,10 @@ export default function SignInWithEmail() {
   const form = useForm<FormData>({
     resolver: zodResolver(formSchema),
     defaultValues: {
+      name: '',
       email: '',
       password: '',
+      agreement: false,
     },
   });
 
@@ -53,30 +68,23 @@ export default function SignInWithEmail() {
     setErrorMessage('');
 
     try {
-      // 使用 API 路由進行登入
-      const response = await fetch('/api/auth/email/login', {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify({
-          email: data.email,
-          password: data.password,
-        }),
-      });
-
-      const result = await response.json();
+      // 調用註冊 API
+      const response = await registerWithEmail(
+        data.email,
+        data.name,
+        data.password,
+      );
 
       // 檢查回應狀態
-      if (result.StatusCode === 200) {
-        // 登入成功，導向首頁
-        router.push('/');
+      if (response.StatusCode === 200) {
+        // 註冊成功，導向驗證頁面
+        router.push('/login-verify');
       } else {
         // 顯示錯誤訊息
-        setErrorMessage(result.msg || '登入失敗，請確認帳號密碼');
+        setErrorMessage(response.msg || '註冊失敗，請稍後再試');
       }
     } catch (error) {
-      console.error('登入失敗:', error);
+      console.error('註冊失敗:', error);
       setErrorMessage('系統錯誤，請稍後再試');
     } finally {
       setSubmitting(false);
@@ -96,7 +104,7 @@ export default function SignInWithEmail() {
             className="mb-6"
           />
           <h1 className="text-center text-2xl font-medium text-neutral-700">
-            歡迎回來!
+            使用 電子郵件 註冊
           </h1>
         </div>
 
@@ -110,9 +118,42 @@ export default function SignInWithEmail() {
         {/* 表單區塊 */}
         <Form {...form}>
           <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-6">
+            {/* 姓名輸入區塊 */}
+            <div className="space-y-2">
+              <div className="text-neutral-700">您的姓名</div>
+              <FormField
+                control={form.control}
+                name="name"
+                render={({ field }) => (
+                  <FormItem className="relative">
+                    <FormControl>
+                      <div className="relative">
+                        <Input
+                          placeholder="請輸入您的姓名"
+                          className="h-14 pl-4 pr-10 rounded-md bg-neutral-50 text-base"
+                          {...field}
+                        />
+                        {field.value && (
+                          <button
+                            type="button"
+                            onClick={() => field.onChange('')}
+                            className="absolute right-3 top-1/2 -translate-y-1/2"
+                            aria-label="清除姓名"
+                          >
+                            <X className="h-6 w-6 text-neutral-400" />
+                          </button>
+                        )}
+                      </div>
+                    </FormControl>
+                    <FormMessage />
+                  </FormItem>
+                )}
+              />
+            </div>
+
             {/* 電子郵件輸入區塊 */}
             <div className="space-y-2">
-              <div className="text-neutral-700">您的電子信箱</div>
+              <div className="text-neutral-700">電子郵件設定</div>
               <FormField
                 control={form.control}
                 name="email"
@@ -121,8 +162,8 @@ export default function SignInWithEmail() {
                     <FormControl>
                       <div className="relative">
                         <Input
-                          placeholder="請輸入您的電子信箱"
-                          className="h-14 pl-4 pr-10 rounded-md bg-gray-50 text-base"
+                          placeholder="請輸入電子郵件地址"
+                          className="h-14 pl-4 pr-10 rounded-md bg-neutral-50 text-base"
                           {...field}
                         />
                         {field.value && (
@@ -145,7 +186,7 @@ export default function SignInWithEmail() {
 
             {/* 密碼輸入區塊 */}
             <div className="space-y-2">
-              <div className="text-neutral-700">您的密碼</div>
+              <div className="text-neutral-700">密碼設定</div>
               <FormField
                 control={form.control}
                 name="password"
@@ -155,8 +196,8 @@ export default function SignInWithEmail() {
                       <div className="relative">
                         <Input
                           type="password"
-                          placeholder="請輸入您的密碼"
-                          className="h-14 pl-4 pr-10 rounded-md bg-gray-50 text-base"
+                          placeholder="請輸入密碼"
+                          className="h-14 pl-4 pr-10 rounded-md bg-neutral-50 text-base"
                           {...field}
                         />
                         {field.value && (
@@ -175,26 +216,50 @@ export default function SignInWithEmail() {
                   </FormItem>
                 )}
               />
+              <div className="text-xs text-neutral-500">
+                密碼需至少8碼，包含至少一個大寫英文字母與一個數字
+              </div>
             </div>
 
-            {/* 忘記密碼連結 */}
+            {/* 同意條款勾選區塊 */}
+            <FormField
+              control={form.control}
+              name="agreement"
+              render={({ field }) => (
+                <FormItem className="flex items-start space-x-3 space-y-0">
+                  <FormControl>
+                    <Checkbox
+                      checked={field.value}
+                      onCheckedChange={field.onChange}
+                      className="h-6 w-6 rounded data-[state=checked]:bg-primary data-[state=checked]:text-white border-neutral-300"
+                    />
+                  </FormControl>
+                  <div className="text-sm text-neutral-700 leading-normal">
+                    我已同意本公司隱私及著作權條款
+                  </div>
+                  <FormMessage />
+                </FormItem>
+              )}
+            />
+
+            {/* 已有帳號連結 */}
             <div className="flex items-center justify-start">
+              <span className="text-neutral-700">已有帳號? </span>
               <Link
-                href="/forgot-password"
-                className="flex items-center text-[#FF4500] hover:underline"
+                href="/signin-email"
+                className="ml-1 text-primary hover:underline"
               >
-                <Lock className="mr-1 h-4 w-4" />
-                忘記密碼
+                登入帳號
               </Link>
             </div>
 
-            {/* 登入按鈕 */}
+            {/* 註冊按鈕 */}
             <Button
               type="submit"
               disabled={submitting}
-              className="w-full h-14 rounded-full bg-[#FF4500] hover:bg-[#FF4500]/90 text-white font-medium text-lg"
+              className="w-full h-14 rounded-full bg-primary hover:bg-primary-50 text-white font-medium text-lg"
             >
-              {submitting ? COMMON_TEXTS.SUBMITTING : '登入'}
+              {submitting ? COMMON_TEXTS.SUBMITTING : '註冊'}
             </Button>
           </form>
         </Form>
